backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and caching
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Data validation
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# AI/ML libraries
openai==1.3.7
anthropic==0.7.7
langchain==0.0.350
llama-index==0.9.13
pinecone-client==2.2.4
sentence-transformers==2.2.2

# HTTP Client
httpx==0.25.2

# Data processing
pandas==2.1.4
numpy==1.25.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
requests==2.31.0

# Background tasks
celery==5.3.4
flower==2.0.1

# Monitoring and logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0
